Refactor route group
