+++ app/(default)/cars/TrendChart.tsx
